backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
mysql-connector-python
pydicom
Pillow
numpy
